requirements.txt
flask
pyzbar
Pillow
requests
